+++ application/src/main/java/com/foodielog/application/user/controller/UserAuthController.java
@@ -4,8 +4,8 @@ import com.foodielog.application.user.dto.UserRequest;
 import com.foodielog.application.user.dto.UserResponse;
 import com.foodielog.application.user.service.OauthUserService;
 import com.foodielog.application.user.service.UserService;
-import com.foodielog.server._core.ExternalAPIUtil.ApiUtils;
-import com.foodielog.server._core.ExternalAPIUtil.CookieUtil;
+import com.foodielog.server._core.util.ApiUtils;
+import com.foodielog.server._core.util.CookieUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
