+++ core/src/main/java/com/foodielog/server/_core/error/exception/Exception400.java
@@ -1,6 +1,6 @@
 package com.foodielog.server._core.error.exception;
 
-import com.foodielog.server._core.ExternalAPIUtil.ApiUtils;
+import com.foodielog.server._core.util.ApiUtils;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
